:sparkles: Adicionar funcionalidade de pontos de recompensa e novas classes de serviço
- **Funcionalidade de Pontos de Recompensa**
  - Implementada a funcionalidade de pontos de recompensa na classe "Cliente".

- **Novas Classes de Serviço**
  - Criadas novas classes de serviço - "ProdutoService" e "PedidoService", com seus respectivos testes.

- **Melhorias na Classe "Pedido"**
  - Realizadas melhorias no cálculo do custo total na classe "Pedido".

+++ src/test/java/diegosneves/ddd/github/entity/PedidoTest.java
@@ -132,10 +132,10 @@ class PedidoTest {
 
     @Test
     void quandoCalcularCustoTotalForChamadoEntaoPrecoDosItensSeraoSomados() {
-        BigDecimal expected = BigDecimal.valueOf(22.0);
+        BigDecimal expected = BigDecimal.valueOf(22);
         BigDecimal actual = this.pedido.calcularCustoTotal();
 
-        assertEquals(expected, actual);
+        assertEquals(expected.doubleValue(), actual.doubleValue());
     }
 
     @Test
